Split subdir.mk into multiple files
Put make rules into their own files, so that changing options or adding
file-specific options on single files does not change any makefile
fragments that are listed as prerequisites of otherwise unaffected files
and unnecessarily require them to be rebuilt. Before, any change to
subdir.mk would cause the whole subdirectory to be rebuilt even when the
rules only changed for a subset of files.

(Removing file-specific options on a single file still causes a full
rebuild through a different mechanism, changing that is outside of the
scope of this commit.)